refactor: ProductProcessing to consolidate stream pipeline and change price to double
- Change `main()` to `static void main(String[] args)` for proper Java entry point
- Move products list initialization inside main method
- Consolidate stream operations: combine filtering, sorting, and mapping into single pipeline
- Remove intermediate variables and debug print statements
- Change price field type from `int` to `double` in Product record
- Update price formatting to use `%.1f` for decimal display

+++ src/main/java/week7/exercise2/Product.java
@@ -1,3 +1,3 @@
 package week7.exercise2;
 
-public record Product(String name, String category, int price, int stockQuantity) {}
+public record Product(String name, String category, double price, int stockQuantity) {}
